Hack sphinx gallery to edit example scripts
This hack would discard anything after the '# DISCARD BELOW' comment in
the example scripts, so an edited version of example scripts are used.
This is useful for removing lines of example scripts that are not part
of the example but are required for desired display of the example
through sphinx. For example, for displaying images that are not
generated by running the example script, some lines should be added to
the end of example script, but those are not relevant to the example and
confuses the user, so it is better to remove them.